Pull into own file and rename to PartialRangeErrorException
